requirements.txt
django==5.1.3
djangorestframework==3.15.2
pip>=24.0.0